requirements.txt
tree-sitter==0.25.2
tree-sitter-c-sharp==0.23.5
qdrant-client==1.12.1
httpx==0.27.2
tqdm==4.66.5
openai>=1.30.0
python-dotenv>=1.0.0
questionary>=2.0.1
